get_file command and on_command_error event completed
The get_file command has been implemented to allow the user to pull files from the Canvas website based off of one parameter, the file's id. As long as the file is in a course the user is currently a part of, they will be able to properly execute this command. The on_command_error event was added last minute to account for any all errors that fall under the category of MissingRequiredArgument. This event works across all commands currently implemented within the bot.