Create/Delete poll commands and timer function
The create polls command was updated to the latest version of the canavsapi python wrapper. The command now works properly and allows the user to create a poll under their account. The delete poll command was added to allow the user to delete any polls they have created by passing the poll's id. A timer function prototype was created to hopefully integrate a notification function into the bot.